MSDKUI-1722: [Android] Component app findings Fixed GuidanceNextManeuverView -- Removed default option from dev app GuidanceNextManeuverView -- added new option without street name in dev app GuidanceNextManeuverView -- hide dot if street or distance view is gone GuidanceNextManeuverView -- Fixed width

+++ MSDKUIKit/MSDKUILib/src/main/java/com/here/msdkui/guidance/GuidanceNextManeuverPresenter.java
@@ -75,7 +75,7 @@ public class GuidanceNextManeuverPresenter extends BaseGuidancePresenter {
             notifyDataChanged(null);
         } else {
             final GuidanceNextManeuverData data = new GuidanceNextManeuverData(getIcon(maneuver),
-                    maneuver.getDistanceFromPreviousManeuver(), getStreet(maneuver));
+                    (long) maneuver.getDistanceFromPreviousManeuver(), getStreet(maneuver));
             if (data.getIconId() == 0 || data.getDistance() < 0 ||
                     data.getDistance() > DistanceFormatterUtil.THOUSAND) {
                 notifyDataChanged(null);
